+++ srecord/memory/walker/stm32.cc
@@ -41,7 +41,7 @@ srecord::memory_walker_stm32::create()
 
 
 void
-srecord::memory_walker_stm32::observe(unsigned long, const void *data,
+srecord::memory_walker_stm32::observe(uint32_t, const void *data,
     int length)
 {
     checksum.nextbuf(data, length);
